+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Support writing to stdout in the pull command and in the sink step when `-` is specified as the db
+- Support reading from stdin in the pull command when `-` is specified as the source
 - Support reading from stdin in generator and html steps when `-` is specified as the source
 - Add a `--def` argument to the `flow` command to provide or override the flow definition
 - Add a `--use-free-ports` argument to the `flow` command